Add HTTPS support for ellis

+++ root/usr/share/clearwater/bin/poll_ellis.sh
@@ -37,10 +37,22 @@
 . /etc/clearwater/config
 http_ip=$(/usr/share/clearwater/bin/bracket_ipv6_address.py $local_ip)
 
-# Send HTTP request and check that the response is "OK".
 http_url=http://$http_ip/ping
-curl -f -g -m 2 -s $http_url 2> /tmp/poll-ellis.sh.stderr.$$ | tee /tmp/poll-ellis.sh.stdout.$$ | head -1 | egrep -q "^OK$"
-rc=$?
+ssl_cert_file=/etc/nginx/ssl/ellis.crt
+ssl_key_file=/etc/nginx/ssl/ellis.key
+
+if [ -e "$ssl_cert_file" ] && [ -e "$ssl_key_file" ]
+then
+  # Send HTTP request and check that a 301 response is sent
+  curl -f -g -m 2 -s $http_url -o /dev/null -w "%{http_code}" 2> /tmp/poll-ellis.sh.stderr.$$ | tee /tmp/poll-ellis.sh.stdout.$$ | egrep -q "^301$"
+  rc=$?
+
+else
+  # Send HTTP request and check that the response is "OK".
+  curl -f -g -m 2 -s $http_url 2> /tmp/poll-ellis.sh.stderr.$$ | tee /tmp/poll-ellis.sh.stdout.$$ | head -1 | egrep -q "^OK$"
+  rc=$?
+
+fi
 
 # Check the return code and log if appropriate.
 if [ $rc != 0 ] ; then
